练习

+++ backend/src/main/java/edu/zxy/apple/condition/vo/AcctListCondVO.java
@@ -1,18 +1,17 @@
 package edu.zxy.apple.condition.vo;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public class AcctListCondVO extends BaseCondVO
 {
-    private List<BigDecimal> acctIdList;
+    private List<Integer> acctIdList;
 
-    public List<BigDecimal> getAcctIdList()
+    public List<Integer> getAcctIdList()
     {
         return acctIdList;
     }
 
-    public void setAcctIdList(List<BigDecimal> acctIdList)
+    public void setAcctIdList(List<Integer> acctIdList)
     {
         this.acctIdList = acctIdList;
     }
